break out api utils
